13.1

+++ Framework-Headers/PrivateFrameworks/CoreSpeech._framework/Objc/CSSPGEndpointAnalyzerDelegate-Protocol.h
@@ -11,7 +11,7 @@
 @protocol CSSPGEndpointAnalyzerDelegate <NSObject>
 
 @optional
-- (void)spgEndpointAnalyzer:(CSSPGEndpointAnalyzer *)arg1 hasSilenceScoreEstimate:(float)arg2;
+- (void)spgEndpointAnalyzer:(CSSPGEndpointAnalyzer *)arg1 hasSilenceScoreEstimate:(double)arg2;
 - (void)spgEndpointAnalyzerDidDetectEndpoint:(CSSPGEndpointAnalyzer *)arg1;
 @end
